backend/app/test/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
passlib
bcrypt==4.0.1
python-multipart
